Intial Outline in README
Defining the initial outline of the content that I borrrowed from the TDD deck taught at ChefConf. Which borrows heavily from the TDD portions of the Intermediate deck.

I re-ordered the outline to present the concepts from the inside-out. This seems like the logical progression in terms of time-investment and resources.

I wrote the initial words on the reasons behind cookbook style and correctness. They are great but they need to be edited.