ChefSpec
Added the sections on reading the error message and re-running specific files
Finished a bit about the coverage
Then talked about the spec_helper and why it is used
After that I dropped a paragraph that points to the documentation in the README. This likely leaves a lot of unanswered questions. Particularly because there are more features to ChefSpec. I think for the purposes of this presentation that we don't need to write this into full detail yet.