Finished Up Test Kitchen, ServerSpec, and Workstation Setup
Rough draft for the remaining setions. For ServerSpec I am relying on the content for RSpec to be reused. Each individual helper will later need to be described or maybe I could come up with a way to generically describe them.

Test Kitchen there is way more that could be said about it. Particularly about passing attributes and other values to the drivers, machines and provisioners.