Scene Breakdown
Alright looking at the big ball of wax I finally started to think about ways to break it down. But the work ahead looks long and ardious -- frought with moments of bikeshedding and worrying about the example. I guess the benefit of the structure is that I can minimize the risk in the training (namely the importance and value of the exercise) and isolate the parts that are known are better known like the introductions and the examination of what the tool or concept that is being covered. The summary sections at the end are short and if they need to be changed if the workshop changes then its not a huge issue.

Ultimately this is starting to feel like the right direction. It also makes it much more manageable when asking someone to give me a hand with an edit or the creation of the exercise.