ChefSpec split apart
This was a harder decision because there feels like a need to extract RSpec out of the ChefSpec part because that is a section we can use again before we talk about ServerSpec (or at least make it optionally available to watch).

So it seemed important to have a scene dedicated to using the Rspec command-line application and one dedicated to the actual language of defining the tests.