Rakefile with rake scene SCENE=X
Found a quick command-line tool on MacOSX called 'ghi' which allows you to easily create github issues through github so I made a quick task that allowed me to create the files specified by the task and then open an issue up on Github.

This is so that I can better track the progress of where I am at and the completion of this attempt at building course material.