Scene 02 Updated
Needed to update the flow to make sure we talk about one new slide that was added to the deck that showed a reminder of the position linting falls into what we are trying to accomplish.

Then the ending objectives needed to be updated to be accurate and grammaticallybetter stated.

Also tried to make the slides more about linting on the introduction and the words. The use of Rubocop is being used to help teach us about linting. Remember we are not the tool -- we are the practice.